Added docs to samples endpoint

+++ services/api/routers/samples.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Request, HTTPException
+from fastapi import APIRouter, Request, HTTPException, Query, Path
 from services.api.schemas.samples import SampleIn, Sample
 from services.api.schemas.annotations import Annotation
 from services.api.schemas import convert_to_objectid
@@ -6,8 +6,23 @@ from services.api.schemas import convert_to_objectid
 router = APIRouter(prefix="/projects/{project_id}/samples", tags=["Samples"])
 
 
-@router.get("", response_model=list[Sample])
-async def get_samples(request: Request, project_id: str, start: int = 0, end: int = None) -> list[Sample]:
+@router.get(
+    "", 
+    response_model=list[Sample],
+    responses = {
+        200: {"description": "Samples have been retrieved successfully."},
+        404: {"description": "Project not found with that ID."},
+    })
+async def get_samples(
+    request: Request, 
+    project_id: str = Path(description="The ID of the project to get samples for."), 
+    start: int = Query(0, description="Index of the first sample you want returned when sorted newest - oldest"), 
+    end: int = Query(None, description="Index of the last sample you want returned when sorted newest - oldest, leave blank to return all entries")
+    ) -> list[Sample]:
+    """
+    Get the full list of samples available for this project.
+    --------------------------------------------------------
+    """
     # Return a list of all samples for this project and info about them    
     project_obj_id = convert_to_objectid(project_id, "projects")
     
@@ -26,13 +41,28 @@ async def get_samples(request: Request, project_id: str, start: int = 0, end: in
     return _samples
 
 
-@router.put("")
-async def add_samples(request: Request, project_id: str, samples: list[SampleIn]):
+@router.put(
+    "",
+    responses = {
+        200: {"description": "Samples have been added successfully, and a list of their IDs has been returned."},
+        404: {"description": "Project not found with that ID."},
+    })
+async def add_samples(
+    request: Request, 
+    samples: list[SampleIn],
+    project_id: str = Path(description="The project ID to associate these samples with.")
+    ):
+    """
+    Add a list of samples (with optional annotations) to this project.
+    ------------------------------------------------------------------
+    """
     # Add samples from the range specified to the project
     # I'm assuming these will be shot/pulse numbers, hence int, but could be unique ID strings instead
     # Depends if for us a 'sample' will always be a shot/pulse, or if it could be a subset eg a single frame of video
     # Do we also want to allow a single value, or list of specific value?
     project_obj_id = convert_to_objectid(project_id, "projects")
+    if not await request.app.state.db_client.get_document_by_id("projects", project_obj_id):
+        raise HTTPException(status_code=404, detail="Project not found with that ID.")
     
     # Remove annotations (if they exist), these will be added later
     all_annotations = [sample.annotations for sample in samples]
@@ -83,14 +113,31 @@ async def add_samples(request: Request, project_id: str, samples: list[SampleIn]
     
     return ids
 
-@router.get("/next", response_model=Sample)
-async def get_next_sample(request: Request, project_id: str) -> Sample:
+@router.get(
+    "/next", 
+    response_model=Sample,
+    responses = {
+        200: {"description": "The next sample to annotate has been returned, according to the project's query strategy."},
+        204: {"description": "No more samples are available to annotate for this project."},
+        409: {"description": "Server is not setup to use the selected project."}
+    }
+    )
+async def get_next_sample(
+    request: Request, 
+    project_id: str = Path(description="The project to return the next sample from.")
+    ) -> Sample:
+    """
+    Get the next sample to annotate for this project, according to query strategy.
+    ------------------------------------------------------------------------------
+    """
     # Return the next sample for human validation for this project
     # Should use the query strategy, which access the database to determine the next sample to annotate
     # This should then be passed in to the /data endpoint to get required data for visualisation
     # And the /annotation endpoint to get initial prediction (if available)
     if not request.app.state.project:
         raise HTTPException(status_code=400, detail="You must setup a project first!")
+    elif request.app.state.project.id != project_id:
+        raise HTTPException(status_code=409, detail="Server is not setup for this project!")        
     try:
         sample = request.app.state.data_pool.query_strategy.get_next_sample()
     except RuntimeError as e:
@@ -98,8 +145,22 @@ async def get_next_sample(request: Request, project_id: str) -> Sample:
     
     return sample
 
-@router.get("/{sample_id}", response_model=Sample)
-async def get_sample(request: Request, project_id: str, sample_id: str) -> Sample:
+@router.get(
+    "/{sample_id}", 
+    response_model=Sample,
+    responses = {
+        200: {"description": "Samples has been returned successfully."},
+        404: {"description": "Either Project or Sample not found with specified ID."},
+    })
+async def get_sample(
+    request: Request, 
+    project_id: str = Path(description="The ID of the project to retrieve a sample from."), 
+    sample_id: str = Path(description="The ID of the sample to retrieve.")
+    ) -> Sample:
+    """
+    Get the specified sample from this project.
+    --------------------------------------------
+    """
     # Get sample with this ID
     project_obj_id = convert_to_objectid(project_id, "projects")
     sample_obj_id = convert_to_objectid(sample_id, "samples")
@@ -118,13 +179,29 @@ async def get_sample(request: Request, project_id: str, sample_id: str) -> Sampl
     return samples[0]
 
 
-@router.delete("/{sample_id}")
-async def remove_sample(request: Request, project_id: str, sample_id: str):
+@router.delete(
+    "/{sample_id}",
+    responses = {
+        200: {"description": "Samples has been returned successfully."},
+        404: {"description": "Sample not found with specified ID within selected project."},
+    })
+async def remove_sample(
+    request: Request,
+    project_id: str = Path(description="The ID of the project to delete a sample from."), 
+    sample_id: str = Path(description="The ID of the sample to delete.")
+    ):
+    """
+    Get the specified sample from this project.
+    --------------------------------------------
+    """
     # Remove samples from the project
     # Dont envisage this actually deleting the data stored about these samples
     # But do we need a separate method for that?
     project_obj_id = convert_to_objectid(project_id, "projects")
     sample_obj_id = convert_to_objectid(sample_id, "samples")
+    
+    if not await request.app.state.db_client.get_document_by_id("projects", project_obj_id):
+        raise HTTPException(status_code=404, detail="Project not found with that ID.")
     
     result = await request.app.state.db_client.delete_filtered_documents(
         collection="projects",
